+++ src/DbLocalizationProvider.Storage.PostgreSQL/ConfigurationContextExtensions.cs
@@ -3,7 +3,6 @@
 
 using System;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.Commands;
 using DbLocalizationProvider.Sync;
 
 namespace DbLocalizationProvider.Storage.PostgreSql
@@ -35,8 +34,6 @@ namespace DbLocalizationProvider.Storage.PostgreSql
 
             context.TypeFactory.ForQuery<UpdateSchema.Command>().SetHandler<SchemaUpdater>();
             context.TypeFactory.ForQuery<SyncResources.Query>().SetHandler<ResourceSynchronizer>();
-
-            context.TypeFactory.ForCommand<RemoveTranslation.Command>().SetHandler<RemoveTranslationHandler>();
 
             return context;
         }
