+++ src/DbLocalizationProvider.Storage.SqlServer/ConfigurationContextExtensions.cs
@@ -3,7 +3,6 @@
 
 using System;
 using DbLocalizationProvider.Abstractions;
-using DbLocalizationProvider.Commands;
 using DbLocalizationProvider.Sync;
 
 namespace DbLocalizationProvider.Storage.SqlServer
@@ -35,8 +34,6 @@ namespace DbLocalizationProvider.Storage.SqlServer
 
             context.TypeFactory.ForQuery<UpdateSchema.Command>().SetHandler<SchemaUpdater>();
             context.TypeFactory.ForQuery<SyncResources.Query>().SetHandler<ResourceSynchronizer>();
-
-            context.TypeFactory.ForCommand<RemoveTranslation.Command>().SetHandler<RemoveTranslationHandler>();
 
             return context;
         }
